+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+## :pencil: Changed
+- Add support for servings other than Peoples (e.g.: quantities)
+- Servings label renamed to "Servings" instead of "People"
+
 ## [1.1.1]
 ### :pencil: Changed
 - Add support for some ingredient volume convertion (Wine).